pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "voxta-twitch-relay"
version = "0.1.0"
description = "A relay app that bridges Twitch chat with the Voxta conversational AI platform via Voxta Gateway."
readme = "README.md"
authors = [{ name = "Dion Labs" }]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "twitchio>=2.8.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    "voxta-gateway>=0.1.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "coverage-badge>=1.1.0",
]
docs = [
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "pymdown-extensions>=10.0",
]

[project.scripts]
voxta-twitch-relay = "voxta_twitch_relay.main:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM", "ARG"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=voxta_twitch_relay --cov-report=term-missing --cov-report=xml --junitxml=junit.xml"

[tool.coverage.run]
source = ["voxta_twitch_relay"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]

[tool.setuptools.packages.find]
exclude = ["site*", "docs*", "tests*"]

[dependency-groups]
dev = [
    "mkdocs-material>=9.7.1",
    "twine>=6.2.0",
]

[project.urls]
Homepage = "https://github.com/dion-labs/voxta-twitch-relay"
Documentation = "https://twitch.voxta.dionlabs.ai"
